requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
asyncpg==0.29.0
aiohttp==3.9.5
alembic==1.13.1
ruff==0.4.4
pyjwt[crypto]==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
